tests/submission/W3C/domhtml/HTMLAnchorElement07.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Anchor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-a-element.html#dom-a-rel" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<a accesskey="g" charset="US-ASCII" coords="0,0,100,100" dir="LTR" href="/support/images/yellow-rectangle.png" hreflang="en" id="Anchor" name="Anchor" rel="GLOSSARY" rev="STYLESHEET" shape="rect" tabindex="22" type="image/png">View Submit Button</a>
</p>


<script>
var domtest = async_test("The rel attribute contains the forward link type.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vrel;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("a");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vrel = testNode.rel;

      assert_equals(vrel,"GLOSSARY");
       
    domtest.done();
}));

</script>
<p>The rel attribute contains the forward link type. 

    Retrieve the rel attribute and examine its value.</p>
<div id="log"></div>
</body></html>
